Conexión con el frontend

+++ app/Notifications/VerifyEmailNotification.php
@@ -14,7 +14,7 @@ class VerifyEmailNotification extends VerifyEmail
     {
         return URL::temporarySignedRoute(
             'api.v1.verification.verify',
-            Carbon::now()->addMinutes(Config::get('auth.verification.expire', 60)),
+            Carbon::now()->addMinutes((int) Config::get('auth.verification.expire', 60)),
             [
                 'id' => $notifiable->getKey(),
                 'hash' => sha1($notifiable->getEmailForVerification()),
